Add InterpreterStorage trait to encapsulate VM req
The users of this library must be agnostic of the internal requirements
of the VM.

This way, a trait encapsulating all these implementations required from
the storage is needed to provide a fully implemented Interpreter.

Following the same pattern, if we need to introduce an ALU-only storage
implementation, we may create the AluStorage trait with the required
signatures.