Add debug feature
The internal debugger of the interpreter will behave as a state machine
with idempotent methods - except for `eval_debugger_state`. This one
will pop the last state and continue the execution of programs.

All debug functionalities are wrapped inside this new feature, and no
additional overhead will occur with the execution outside the debug
environment.