Add debug feature (#3)
* Merge executors module

* Add debug feature

The internal debugger of the interpreter will behave as a state machine
with idempotent methods - except for `eval_debugger_state`. This one
will pop the last state and continue the execution of programs.

All debug functionalities are wrapped inside this new feature, and no
additional overhead will occur with the execution outside the debug
environment.

* Add debug test case for tx script

* Move all contract addr logic to debugger

Currently, when the contract address is not defined, the zeroed address
is being used.

All implementations of this logic should be internal to the debugger
module.