Add MerkleStorage trait to data
The merkle storage trait will extend the storage functionality, adding
the ability to calculate the merkle root.

The contract code root assumes a small in-memory merkle tree is used as
backend to calculate the contract id.

This commit also included the improvement over the instructions
execution that the opcodes are enforced by design to have consistent
register references (e.g. they only stored in 6 bits - hence they can't
overflow the 64 available registers). This means that range check for
read-only registers is not required.