Add `from_slice_unchecked` to tx types (#27)
Very often the VM needs to convert between arbitrary memory slices into
the tx types. The memory bounds of the VM memory are logically verified
and this is a safe operation.

This pattern of expecting the converted type from a slice via `TryFrom`
is recurrent and should be upgraded into a native function of the type
itself.