Refactor `execute` to parse Word in pairs
With the introduction of `fuel_asm::Instruction`, we can now parse pairs
of instructions from words, reducing the execution overhead compared to
when we parse the opcodes individually from half-words.

This allows also the static gas calculation for most opcodes. For the
ones with variable cost (e.g. MCL), a monad concept was introduced.

The variable gas cost opcodes should be described in `term + F(x)` form,
where `term` is the constant part and `F` should take `GasUnit` to
define a simplified arithmetic operation with the runtime register
value.