Add `Instruction::into_inner` (#27)
An instruction, after parsed, will have all its attributes exposed to
the VM. Its convenient to have a canonical `into_inner` so the
underlying fields can be read.

The exposed fields are:

- OpcodeRepr
- RegisterId $ra
- RegisterId $rb
- RegisterId $rc
- RegisterId $rd
- Word immediate value

To achieve that, `OpcodeRepr` must be refactored before its exposed as
public. We should cover all variants of a byte in order to perform
transparent safe conversions from/to `OpcodeRepr`/`u8` in the internal
representation.