Add `Backtrace` error struct
Backtrace will be responsible to provide the required data to debug the
souce of an error.

Since its creation cost isn't cheap, it will not be provided by default.
The user will need to create a backtrace instance from the VM instance
that generated the error, combined with the error variant itself.