Update contract id to use celestia specs BMT
The binary merkle tree of the specs points to the one specified in
celestia-specs. They are protected against second-preimage attacks by
prefixing the inputs with 0x00 for leaves, or 0x01 for nodes.

This commit introduces this protection.

Also, as window of opportunity, the deprecated since 1.56 `use TryFrom`
was removed from the code.