Add `Opcode::BAL` implementation
The opcode implementation brings the ability to inspect the balance of a
given color+contract tuple.

This allows the test case `mint_burn` to be moved to the integrated
scope since before this opcode, there was no way to fetch the balance of
the tuple without inspecting the internals of the interpreter.

Follows https://github.com/FuelLabs/fuel-specs/pull/226