Add `secp256k1` signature for `Input::Coin` (#57)
Add recoverable signature to input coins.

The signed message is the transaction hash. Since it can be signed only
after the transaction is built, the coin input can only be appended
before it is signed since it will define the transaction hash.

This created the need to include a `TransactionBuilder` that will accept
the arguments of the transaction, and sign the coin inputs after the
transaction id is ready.

Also, the signed transaction cannot contain the witnesses.

These changes are breaking regarding the tests; they now should use the
builder structure when input signatures are to be validated.