Add static secp256k1 for sig w/ benches (#22)
The secp256k1 provider shouldn't be instantiated on every request.
Instead, it should be instantiated once, with a lazy strategy, and
made available for signature requests.

This commit introduces `lazy_static` to achieve that. Lazy is still
experimental in `std`, so we need to wait for its stabilization before
requiring it.

In order to increase the robustness of the secp provider, we introduce
`rand-std` for `secp256k1`. This feature will boost the entropy of the
blinders, further protecting the implementation from side-channel
attacks. This will not remove the deterministic nonce requirement, as
defined in the RFC-6979 and the fuel crypto specs.

Include benchmarks comparing `fuel-crypto`, `secp256k1` and `k256`

Resolves #3