Update instruction executors to not dupl code
The instruction executors are duplicated into different implementations
that either support the interpreter storage or don't. The latter is used
for predicate verification since contract opcodes, the only ones that
performs storage calls, aren't allowed.

This duplication will lead to harder long-term maintenance and should be
avoided.

The alternative introduced in this PR is to have a concrete storage
implementation that will deny any storage call, and introduce a check
for each instruction to evaluate if its predicate verification and if
the opcode is allowed for predicates.

This is sub-optimal because we introduce an extra branching per VM
cycle, but it can easily be refactored later to perform a couple of
bitwise operations after fuel-asm #68 is done.

Closes #133