Implement GM::GetVerifyingPredicate metadata instr
The context of the execution should contain more info such as the index
of the current predicate so we decouple that from the vm initialization.

Also, its worthy to create a type that will encapsulate the behavior of
the runtime predicates since it needs to store not only the program to
be executed - so it can't be represented just as a memory slice - but
the original input index that created the predicate.

The input index is not required for the predicate execution since the
program memory offset is a constant function of the consensus
parameters, but it might be consumed by the user - as in the specs.