Add traceable bug id to unexpected behavior (#189)
Some states of the VM are ultimately invalid and contradict the
specifications constraints. These maps to implementation errors that
must be fixed.

In order to make it easier for the user to spot these errors, we must
have a reliable reporting system.

As a first step, its important to map the origin of the error. This
commit introduces `Bug`, that will provide meaningful descriptions to
circumstances that violates the specifications constraints, and `BugId`,
that is guaranteed by tests to map to a specic - and unique - location
of the code.