Add `Interpreter::prepare_call` (#215)
This commit introduces a function to allow users to simulate a call
operation without executing the loaded code/context.

This operation is desirable for bug reproduction or isolated analysis of
instruction executions. Prior to this commit, its impossible to execute
contract instructions in isolation.

This function is expected by the benchmark so an analysis can be done of
the contract operations without the need to correct the results with the
call overhead.