Add `InterpreterStorage::deploy_contract`
Consumers of the VM might want to replicate the contract deployment also
to emulate the behavior of storages.

Currently, the only way to do that is via a `Create` transaction, but
then we might want to facilitate this operation without creating the
whole transaction requirements in order to allow complex operations such
as benchmarking.

This commit introduces the ability to add a contract into an interpreter
storage using the same strategy that transaction execution would use. It
is added as part of the storage because the responsibility for the
implementation is a concrete aspect of the database backend.